Remove code duplication
Refactor into common test class
